+++ packages/orq-rc/RELEASES.md
@@ -8289,3 +8289,13 @@ Based on:
 - [typescript v4.8.0-rc.13] packages/orq-rc
 ### Releases
 - [NPM v4.8.0-rc.13] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.13 - packages/orq-rc
+
+## 2026-04-14 09:38:53
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.761.4 (2.879.13) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.8.0-rc.14] packages/orq-rc
+### Releases
+- [NPM v4.8.0-rc.14] https://www.npmjs.com/package/@orq-ai/node/v/4.8.0-rc.14 - packages/orq-rc